+++ lib/audiodev/AudioSubmix.cpp
@@ -101,7 +101,7 @@ constexpr T ClampInt(float in) {
 
     if (in < MIN)
       return MIN;
-    else if (in > MAX)
+    else if (in > static_cast<float>(MAX))
       return MAX;
     else
       return in;
